pyproject.toml
[project]
name = "pipeline-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==3.1.7",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
]
